Fix bugs and add to readme
last update was broken where things weren't folding flat and right
hand rule wasn't respected.

This fixes things so that the axes are correct and we properly
calculate a new rotation `origami_unfold_angle` which lays two
adjacent faces flat and non-overlapping. The method is a bit cheesy
but works...