add shadow faces work
I've been thinking about how to actually model origami for a bit and
thought this was a step in the right direction. It kinda is but isn't
really necessary so backing it out. But it is kinda useful so leaving it
in this branch.

A "shadow" face is added for every edge that didn't get included in the
spanning tree (ie. where a cut was introduced)

I thought this might be useful for specifying what fold angle you want
that cut edge to be at (and it might still be?) but it actually seems
redundant since the target constraint is more simply that the edge --
and really just the vert that isn't shared -- gets folded to the same
place.