try out jacobian scaling w/ integration
to go from unfold to folded state, we need to get positions of verts all
along the way. Instead of snapping the verts into agreement at
timepoints, we can instead integrate

To get a good derivative, take the jacobian of apply_rotations so that
we know how each angle effects each vert position. We then need to find
a scaling between the angles s.t. the total effect of derivs is the
same between the corresponding verts. Once we have that, can step in
that direction. We multiply by the sign of the target angles so that we
always step in the right direction